feat(create): move Vitest setup to @mirta/testing
