+++ packages/mirta-testing/rollup.config.mjs
@@ -1,0 +1,8 @@
+import { definePackageConfig } from '@mirta/rollup'
+
+export default definePackageConfig({
+  input: [
+    'src/index.ts',
+    'src/setup/global.ts',
+  ],
+})
